BSERVDEV-14785: Pin AMI baseimage
The Amazon Linux AMI, which is the base AMI for our AMI, has been
updated to use a newer Linux kernel. Unfortunately this kernel is not
supported by the currently available version of ZFS on Linux.
In order to maintain compatibility, this change pins the version of the
base AMI and restricts the `yum update` to only security patches.

Once v6.5.10 of ZoL has been released and made available in the RPM,
all the version pins can be removed.